Création du composant de select pour le temps de cuisson de la recette pour simplifier le composant FormCreateRecipe
